refactor details card
